<!-- Ensure that raw HTML is properly formatted -->
<div style="width:100%; height:150px;border-width:4px;border-style:solid;padding-top:25px;
    border-color:#000;border-radius:10px;text-align:center;background-color:#99DDFF;
    color:#000">
    <h3 style="color: black;">Star us on GitHub!</h3>
    <a class="github-button" href="https://github.com/LAMPSPUC/SARIMAX.jl" data-icon="octicon-star" data-size="large" data-show-count="true" aria-label="Star LAMPSPUC/SARIMAX.jl on GitHub" style="margin:auto">Star</a>
    <script async defer src="https://buttons.github.io/buttons.js"></script>
</div>

Sarimax.jl

Installation

This package is registered so you can simply add it using Julia's Pkg manager:

using Pkg
Pkg.add("SARIMAX")

Auto SARIMA implementation

Main.Sarimax.auto — Function

auto(
    y::TimeArray;
    exog::Union{TimeArray,Nothing}=nothing,
    seasonality::Int=1,
    d::Int = -1,
    D::Int = -1,
    maxp::Int = 5,
    maxd::Int = 2,
    maxq::Int = 5,
    maxP::Int = 2,
    maxD::Int = 1,
    maxQ::Int = 2,
    maxOrder::Int = 5,
    informationCriteria::String = "aicc",
    allowMean:Union{Bool,Nothing} = nothing,
    allowDrift::Union{Bool,Nothing} = nothing,
    integrationTest::String = "kpss",
    seasonalIntegrationTest::String = "seas",
    objectiveFunction::String = "mse",
    assertStationarity::Bool = true,
    assertInvertibility::Bool = true,
    showLogs::Bool = false,
    outlierDetection::Bool = false
    searchMethod::String = "stepwise"
)

Automatically fits the best SARIMA model according to the specified parameters.

Arguments

y::TimeArray: The time series data.

exog::Union{TimeArray,Nothing}: Optional exogenous variables. If Nothing, no exogenous variables are used.

seasonality::Int: The seasonality period. Default is 1 (non-seasonal).

d::Int: The degree of differencing for the non-seasonal part. Default is -1 (auto-select).

D::Int: The degree of differencing for the seasonal part. Default is -1 (auto-select).

maxp::Int: The maximum autoregressive order for the non-seasonal part. Default is 5.

maxd::Int: The maximum integration order for the non-seasonal part. Default is 2.

maxq::Int: The maximum moving average order for the non-seasonal part. Default is 5.

maxP::Int: The maximum autoregressive order for the seasonal part. Default is 2.

maxD::Int: The maximum integration order for the seasonal part. Default is 1.

maxQ::Int: The maximum moving average order for the seasonal part. Default is 2.

maxOrder::Int: The maximum order for the non-seasonal part. Default is 5.

informationCriteria::String: The information criteria to be used for model selection. Options are "aic", "aicc", or "bic". Default is "aicc".

allowMean::Union{Bool,Nothing}: Whether to include a mean term in the model. Default is nothing.

allowDrift::Union{Bool,Nothing}: Whether to include a drift term in the model. Default is nothing.

integrationTest::String: The integration test to be used for determining the non-seasonal integration order. Default is "kpss".

seasonalIntegrationTest::String: The integration test to be used for determining the seasonal integration order. Default is "seas".

objectiveFunction::String: The objective function to be used for model selection. Options are "mse", "ml", or "bilevel". Default is "mse".

assertStationarity::Bool: Whether to assert stationarity of the fitted model. Default is true.

assertInvertibility::Bool: Whether to assert invertibility of the fitted model. Default is true.

showLogs::Bool: Whether to suppress output. Default is false.

outlierDetection::Bool: Whether to perform outlier detection. Default is false.

searchMethod::String = "stepwise"

References

Hyndman, RJ and Khandakar. "Automatic time series forecasting: The forecast package for R." Journal of Statistical Software, 26(3), 2008.